Fix title entering
Need to track down why this was actually causing a problem
but prior to this fix hitting enter to submit todos would work,
but sometimes the title would be blank until refreshed.

I was only using 0 in the boundPropertyPath as a shortcut
for testing for completedItems, but count() works fine
and is probably more explicit anyway.