Change clear committed form to div
I'd added this anyway to give us something to hide easily
using the dynamic-element classList bindings, making it a form
only complicated things.

Now if it was a submit and not a button I might bother to
handle the actual submission, but seeing as it's only a button
I'm deferring to the given template.